+++ contracts/factory/schema/wallet_factory_execute_msg.json
@@ -36,7 +36,7 @@
           ],
           "properties": {
             "migration_msg": {
-              "$ref": "#/definitions/ProxyMigrationMsg"
+              "$ref": "#/definitions/ProxyMigrationTxMsg"
             },
             "wallet_address": {
               "$ref": "#/definitions/WalletAddr"
@@ -53,6 +53,28 @@
       ],
       "properties": {
         "update_proxy_code_id": {
+          "type": "object",
+          "required": [
+            "new_code_id"
+          ],
+          "properties": {
+            "new_code_id": {
+              "type": "integer",
+              "format": "uint64",
+              "minimum": 0.0
+            }
+          }
+        }
+      },
+      "additionalProperties": false
+    },
+    {
+      "type": "object",
+      "required": [
+        "update_proxy_multisig_code_id"
+      ],
+      "properties": {
+        "update_proxy_multisig_code_id": {
           "type": "object",
           "required": [
             "new_code_id"
@@ -173,7 +195,7 @@
         }
       }
     },
-    "ProxyMigrationMsg": {
+    "ProxyMigrationTxMsg": {
       "oneOf": [
         {
           "type": "object",
