+++ contracts/factory/schema/wallet_factory_query_msg.json
@@ -1,6 +1,6 @@
 {
   "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "QueryMsg",
+  "title": "WalletFactoryQueryMsg",
   "oneOf": [
     {
       "description": "Shows all the proxy wallet address Returns WalletListResponse",
@@ -22,6 +22,18 @@
       ],
       "properties": {
         "proxy_code_id": {
+          "type": "object"
+        }
+      },
+      "additionalProperties": false
+    },
+    {
+      "type": "object",
+      "required": [
+        "multisig_code_id"
+      ],
+      "properties": {
+        "multisig_code_id": {
           "type": "object"
         }
       },
