--- contracts/proxy/schema/canonical_addr.json
@@ -1,15 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "CanonicalAddr",
-  "allOf": [
-    {
-      "$ref": "#/definitions/Binary"
-    }
-  ],
-  "definitions": {
-    "Binary": {
-      "description": "Binary is a wrapper around Vec<u8> to add base64 de/serialization with serde. It also adds some helper methods to help encode inline.\n\nThis is only needed as serde-json-{core,wasm} has a horrible encoding for Vec<u8>",
-      "type": "string"
-    }
-  }
-}
