+++ contracts/proxy/schema/execute_msg_for__empty.json
@@ -79,9 +79,81 @@
         }
       },
       "additionalProperties": false
+    },
+    {
+      "type": "object",
+      "required": [
+        "add_relayer"
+      ],
+      "properties": {
+        "add_relayer": {
+          "type": "object",
+          "required": [
+            "new_relayer_address"
+          ],
+          "properties": {
+            "new_relayer_address": {
+              "$ref": "#/definitions/Addr"
+            }
+          }
+        }
+      },
+      "additionalProperties": false
+    },
+    {
+      "type": "object",
+      "required": [
+        "remove_relayer"
+      ],
+      "properties": {
+        "remove_relayer": {
+          "type": "object",
+          "required": [
+            "relayer_address"
+          ],
+          "properties": {
+            "relayer_address": {
+              "$ref": "#/definitions/Addr"
+            }
+          }
+        }
+      },
+      "additionalProperties": false
+    },
+    {
+      "type": "object",
+      "required": [
+        "update_guardians"
+      ],
+      "properties": {
+        "update_guardians": {
+          "type": "object",
+          "required": [
+            "guardians"
+          ],
+          "properties": {
+            "guardians": {
+              "$ref": "#/definitions/Guardians"
+            },
+            "new_multisig_code_id": {
+              "type": [
+                "integer",
+                "null"
+              ],
+              "format": "uint64",
+              "minimum": 0.0
+            }
+          }
+        }
+      },
+      "additionalProperties": false
     }
   ],
   "definitions": {
+    "Addr": {
+      "description": "A human readable address.\n\nIn Cosmos, this is typically bech32 encoded. But for multi-chain smart contracts no assumptions should be made other than being UTF-8 encoded and of reasonable length.\n\nThis type represents a validated address. It can be created in the following ways 1. Use `Addr::unchecked(input)` 2. Use `let checked: Addr = deps.api.addr_validate(input)?` 3. Use `let checked: Addr = deps.api.addr_humanize(canonical_addr)?` 4. Deserialize from JSON. This must only be done from JSON that was validated before such as a contract's state. `Addr` must not be used in messages sent by the user because this would result in unvalidated instances.\n\nThis type is immutable. If you really need to mutate it (Really? Are you sure?), create a mutable copy using `let mut mutable = Addr::to_string()` and operate on that `String` instance.",
+      "type": "string"
+    },
     "BankMsg": {
       "description": "The message types of the bank module.\n\nSee https://github.com/cosmos/cosmos-sdk/blob/v0.40.0/proto/cosmos/bank/v1beta1/tx.proto",
       "oneOf": [
@@ -201,6 +273,53 @@
     "Empty": {
       "description": "An empty struct that serves as a placeholder in different places, such as contracts that don't set a custom message.\n\nIt is designed to be expressable in correct JSON and JSON Schema but contains no meaningful data. Previously we used enums without cases, but those cannot represented as valid JSON Schema (https://github.com/CosmWasm/cosmwasm/issues/451)",
       "type": "object"
+    },
+    "Guardians": {
+      "type": "object",
+      "required": [
+        "addresses"
+      ],
+      "properties": {
+        "addresses": {
+          "description": "A List of keys can act as guardian for",
+          "type": "array",
+          "items": {
+            "type": "string"
+          }
+        },
+        "guardians_multisig": {
+          "description": "Whether multisig option for guardians is enabled",
+          "anyOf": [
+            {
+              "$ref": "#/definitions/MultiSig"
+            },
+            {
+              "type": "null"
+            }
+          ]
+        }
+      }
+    },
+    "MultiSig": {
+      "type": "object",
+      "required": [
+        "multisig_initial_funds",
+        "threshold_absolute_count"
+      ],
+      "properties": {
+        "multisig_initial_funds": {
+          "type": "array",
+          "items": {
+            "$ref": "#/definitions/Coin"
+          }
+        },
+        "threshold_absolute_count": {
+          "description": "Only Fixed multisig is supported in this version",
+          "type": "integer",
+          "format": "uint64",
+          "minimum": 0.0
+        }
+      }
     },
     "RelayTransaction": {
       "type": "object",
