+++ contracts/proxy/src/lib.rs
@@ -1,6 +1,8 @@
 pub mod contract;
 mod error;
+pub mod helpers;
 pub mod msg;
 pub mod state;
 
-pub use crate::error::ContractError;
+#[cfg(test)]
+mod tests;
